AOS-350: move the is published checks
